# speed up sample_polar