@bart1 2024-07-17
The same can be done to `integrate_to_ppi` saving about 25% of time. There is room for improvement still there as `as_Spatial` still takes 25% of the time:
``` r
require(bioRad)
#> Loading required package: bioRad
#> Welcome to bioRad version 0.7.3.9000
#> Attempting to load MistNet from: /home/bart/R/x86_64-pc-linux-gnu-library/4.4/vol2birdR/lib 
#> MistNet successfully initialized.
#> using vol2birdR version 1.0.2 (MistNet installed)
template_raster <- raster::raster(ncol=300, nrow=300,
  raster::extent(12, 13, 56, 57),
  crs = sp::CRS("+proj=longlat")
)
pvolfile <- system.file("extdata", "volume.h5", package = "bioRad")

data(example_vp)
pvol <- read_pvolfile(pvolfile)

bench::mark(ppi <- integrate_to_ppi(pvol, example_vp, raster = template_raster), iterations = 5)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 1 × 6
#>   expression                             min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                           <bch> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 ppi <- integrate_to_ppi(pvol, examp…  4.5s  4.95s     0.201     618MB     3.58
unloadNamespace("bioRad")
devtools::load_all("~/bioRad")
#> ℹ Loading bioRad
#> Welcome to bioRad version 0.7.3.9000
#> using vol2birdR version 1.0.2 (MistNet installed)
bench::mark(ppi2 <- integrate_to_ppi(pvol, example_vp, raster = template_raster), iterations = 5)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 1 × 6
#>   expression                             min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                           <bch> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 ppi2 <- integrate_to_ppi(pvol, exam… 2.94s  3.39s     0.301     584MB     3.92
all.equal(ppi, ppi2)
#> [1] TRUE
```

<sup>Created on 2024-07-17 with [reprex v2.1.0](https://reprex.tidyverse.org)</sup>